Changed default paths in align_files.m and made it only run for unpocessed files

+++ src/align_files.m
@@ -1,5 +1,5 @@
-input_dir = '/home/przemek/neurodata/ca_mecp2_culture/tmp';
-out_dir = '/home/przemek/neurodata/ca_mecp2_culture/Mecp2_cultures_aligned2';
+input_dir = '/home/przemek/neurodata/ca_mecp2_culture/2017-06-07/raw';
+out_dir = '/home/przemek/neurodata/ca_mecp2_culture/Mecp2_cultures_aligned';
 
 inputfiles = listfiles(input_dir, '.*tif');
 
@@ -11,9 +11,11 @@ for file_index=1:numel(inputfiles)
     r = rigidAlign(fname);
     tiff_len = numel(imfinfo(fname));
     output_f = [out_dir filesep filename '_aligned' ext];
-    for i=1:tiff_len
-        data = imread(fname,i);
-        data_t = circshift(data, -r.T(i,:));
-        imwrite(data_t,output_f,'WriteMode','append');
+    if ~exist(output_f, 'file')
+        for i=1:tiff_len
+            data = imread(fname,i);
+            data_t = circshift(data, -r.T(i,:));
+            imwrite(data_t,output_f,'WriteMode','append');
+        end
     end
 end
